Move helpers and schema to helpers folder
