+++ transparency-in-coverage/python/processors/helpers/schema.py
